+++ CHANGELOG.md
@@ -62,6 +62,8 @@
 * **\[core\]** Introduce the `@PreUnmount` annotation that can be used to annotate a method that is invoked
   before a component is unmounted or detatched. This replaces the ability to override the
   `componentWillUnmount()` lifecycle method.
+* **\[core\]** Remove the `React.cloneElement(...)` method as it has never been used by any React4j application
+  and if it was used it could be reimplemented in java rather than relying on javascript extern.
 
 ### [v0.108](https://github.com/react4j/react4j/tree/v0.108) (2018-11-22)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.107...v0.108)